+++ gallery_dl/extractor/__init__.py
@@ -106,7 +106,6 @@ modules = [
     "mangadex",
     "mangafox",
     "mangahere",
-    "mangakakalot",
     "manganelo",
     "mangapark",
     "mangaread",