+++ gallery_dl/extractor/chevereto.py
@@ -78,6 +78,8 @@ class CheveretoImageExtractor(CheveretoExtractor):
             "id"   : self.path.rpartition(".")[2],
             "url"  : url,
             "album": text.extr(extr("Added to <a", "/a>"), ">", "<"),
+            "date" : text.parse_datetime(extr(
+                '<span title="', '"'), "%Y-%m-%d %H:%M:%S"),
             "user" : extr('username: "', '"'),
         }